refactor: align project with Agents guidelines
