+++ README.md
@@ -1,1 +1,20 @@
-# node-boilerplate
+# Node Boilerplate
+
+This repository demonstrates a Clean Architecture setup for a Node.js + TypeScript service.
+
+## Getting Started
+
+Install dependencies and run the tests:
+
+```bash
+yarn install
+yarn test
+```
+
+Run the linter:
+
+```bash
+yarn lint
+```
+
+The HTTP API is documented in `src/contracts/service.yaml`.
