+++ jest/stop-integration.ts
@@ -1,6 +1,8 @@
 import { MongoMemoryReplSet } from 'mongodb-memory-server';
 
 export default async function globalTeardown() {
-  const instance: MongoMemoryReplSet = (global as any).__MONGOINSTANCE;
+  const instance: MongoMemoryReplSet = (
+    global as unknown as { __MONGOINSTANCE: MongoMemoryReplSet }
+  ).__MONGOINSTANCE;
   await instance.stop();
 }
