+++ package.json
@@ -4,9 +4,9 @@
   "version": "1.0.0",
   "main": "index.js",
   "scripts": {
-    "dev": "ts-node-dev --env-file=.env src/app.ts",
+    "dev": "ts-node-dev --env-file=.env src/main.ts",
     "build": "tsc -p .;yarn run copy-essentials",
-    "start": "node dist/src/app.js",
+    "start": "node dist/src/main.js",
     "copy-essentials": "mkdir -p dist/src/contracts && cp src/contracts/*.yaml dist/src/contracts",
     "clean": "rm -rf dist",
     "prettier": "prettier --cache --write src/**/*.ts",
@@ -32,9 +32,9 @@
     "@types/node": "^22.10.1",
     "@types/supertest": "^6.0.2",
     "eslint": "^9.16.0",
-    "globals": "^15.12.0",
     "eslint-config-prettier": "^10.0.1",
     "eslint-plugin-prettier": "^5.2.3",
+    "globals": "^15.12.0",
     "husky": "^9.1.7",
     "jest": "^29.7.0",
     "mongodb-memory-server": "^10.1.2",
