+++ src/infrastructure/config/factories/user.controller.factory.ts
@@ -1,5 +1,5 @@
-import { UserController } from '../../application/controllers/user.controller';
-import { IController } from '../../domain/server/interfaces/IController';
+import { UserController } from '../../../interfaces/http/controllers/user.controller';
+import { IController } from '../../../interfaces/http/controllers/IController';
 import { UserServiceFactory } from './user.service.factory';
 
 export class UserControllerFactory {
