+++ src/infrastructure/config/factories/user.service.factory.ts
@@ -1,6 +1,6 @@
-import { UserService } from '../../domain/user/service/user.service';
-import { UserRepositoryRead } from '../../infrastructure/repository/user/user.repository.read';
-import { UserRepositoryWrite } from '../../infrastructure/repository/user/user.repository.write';
+import { UserService } from '../../../domain/user/service/user.service';
+import { UserRepositoryRead } from '../../repository/user/user.repository.read';
+import { UserRepositoryWrite } from '../../repository/user/user.repository.write';
 
 export class UserServiceFactory {
   static create() {
