+++ src/interfaces/http/server.ts
@@ -7,7 +7,7 @@ import express, {
 } from 'express';
 import { ContextAsyncHooks, Logger } from 'traceability';
 import { Server as httpServer } from 'http';
-import { IController } from './interfaces/IController';
+import { IController } from './controllers/IController';
 import mongoose from 'mongoose';
 import * as OpenApiValidator from 'express-openapi-validator';
 import helmet from 'helmet';
